requirements.txt
boto3==1.18.21
easydict==1.9
gensim==4.0.1
h5py==3.3.0
numpy==1.21.1
paddlepaddle-gpu==2.1.2.post101
pandas==1.3.1
tqdm==4.62.0
h5py==3.3.0
